Prepack Analytics + slight changes
Prepack has been added to analytics, and is now able to be checked
alongside other data, on its own, or omitted by the user. Other changes
include minor differences to the appearance of the form, the ability to
force multiple / single bar chart display, and a message that becomes
visible when data collection methods are changed, to tell the user to
click the Load Data button when they're ready.