Fixing crazy averages
So now we no longer have to worry about averages that are created by
crazed speeds. We have a method of cutting off rows from the chart if
they haven't picked / packed [X amount] compared to what people who pick
/ pack are actually managing to do.

Totals of actual orders are looked at, the largest one is kept as a
reference. If people pick / pack too low a percentage in comparison,
they're dropped from the chart. That percentage is editable. This means
that when someone up in the office goes downstairs and scans 3 things at
packing in the space of 10 seconds, it won't matter because 3 is nowhere
close to what people are actually doing, but everyone else down there
will have somewhere much closer.

The default percentage is 2.