Major design overhaul
So the primary stuff is done, we have a base for our new setup. Employee
selection is available, which is lovely.
We're splitting between single day info view and multiple day info view.
We're also adding a new control for the day view to allow specific
by-the-hour work view. Perhaps we can make it modifiable! Maybe.